selection: move to pkg/factory
