+++ internal/test/helper.go
@@ -5,8 +5,8 @@ import (
 	"math/rand"
 	"testing"
 
-	"github.com/aurelien-rainone/evolve/factory"
 	"github.com/aurelien-rainone/evolve/pkg/api"
+	"github.com/aurelien-rainone/evolve/pkg/factory"
 	"github.com/stretchr/testify/assert"
 )
