+++ pkg/factory/bitstring.go
