+++ pkg/factory/bitstring_test.go
