+++ pkg/factory/string.go
