+++ pkg/factory/string_test.go
